Bug Fixes and Lighting Update
Fixed the bug with the lock and key by making a trigger box event for the lock needing a condition for the key to need the Interaction layer mask of "KeyLayer".(For more details ask me (Matthew))

We also fixed the shape of the arrows making them exact and not lob sided.

All lights were spread across the level.

Fixed center room opening it up by reference to the images found of what the room should look like.

Added arrows all over the level and added in trigger boxes and conditions which will toggle Phase 1 2 3 and 4 arrows depending on which trigger box is overlapped however I need to add into this code so that it only triggers the trigger boxes once per overlap as right now the overlap event will happen multiple times if the player overlaps again and again.