Grapple Merge
Grapple gun has been merged and added to level 2 also parts of the puzzle mechanic for the book shelves puzzle has been added in along with multiple models made by myself for the book shelves, books and the new lighting has been added. Also like with skeletal build I have fixed the lights being too many in 1 object to being just 1 light so that the game doesnt lag while walking around.

Fixed minor grapple gun bug making it so that there is a 5 second delay and then it will stop the player from being pulled towards the point of the plungers contact.

Also minor fixed have been done for the ground and made it so that say your on level 2, the open area for the middle will be able to be walked on (turned to glass) but above will still be open.

- Matt