Final Fixes
Fixed a bug for all targets not being attachable due to :is trigger being on. Also added a few more targets to floor 2.

Had to revert the key on floor 1 as it wouldnt slot into the chests so we just have capsules.

Fixed footsteps audio missing bug.

Removed sounds for items being picked as the scrip wouldnt have any cooldown and the collision would make it make the sounds too frequently.

Tested UIs for level 1 and 2 and they seem good with no errors.