Ninja Star + Target
Throwable Ninja Star and a target that it can get stuck are implemented. Ninja Star's weight may need tweaking as it can feel quite light sometimes also they get stuck in the target on purpose. Not sure what we want to do with them so it's just a throwable star that sticks to the target.